requirements.txt
setuptools:
Django
djangorestframework
djoser




 